(PE-19203) Only generate 1.0 pe.conf in 20171
...and move pe.conf generation based on meep_schema_version flag into
2017.2 which is the first version where we'll be testing with variable
pe-modules packages.
