+++ lib/beaker-answers/versions/version20171.rb
@@ -10,12 +10,5 @@ module BeakerAnswers
       /\A2017\.1/
     end
 
-    # This used to generate the profile host parameters, but now generates a MEEP
-    # 2.0 node_roles hash mapping roles -> node certs based on the same host and
-    # role information.
-    def hiera_host_config
-      pe_conf = BeakerAnswers::PeConf.new(@hosts, @options[:meep_schema_version])
-      pe_conf.configuration_hash
-    end
   end
 end
